+++ config/services.yaml
@@ -128,7 +128,7 @@ services:
                 - 'App\Entity\MachineProvider'
 
     App\Services\ServiceStatusInspector\ServiceStatusInspectorInterface:
-        class: App\Services\ServiceStatusInspector\GenericServiceStatusInspector
+        class: App\Services\ServiceStatusInspector\ServiceStatusInspector
         arguments:
             $componentInspectors:
                 database: '@App\Services\ServiceStatusInspector\DatabaseInspector'
@@ -136,13 +136,6 @@ services:
                 machine_provider_digital_ocean: '@App\Services\ServiceStatusInspector\DigitalOceanMachineProviderInspector'
             $exceptionHandlers:
                 - '@App\Services\ServiceStatusInspector\LoggingExceptionHandler'
-
-    App\Services\ServiceStatusInspector\ServiceStatusInspector:
-        arguments:
-            $componentInspectors:
-                database: '@App\Services\ServiceStatusInspector\DatabaseInspector'
-                message_queue: '@App\Services\ServiceStatusInspector\MessageQueueInspector'
-                machine_provider_digital_ocean: '@App\Services\ServiceStatusInspector\DigitalOceanMachineProviderInspector'
 
     SmartAssert\DigitalOceanDropletConfiguration\Factory:
         arguments:
