+++ tests/Functional/Services/ServiceStatusInspector/ServiceStatusInspectorTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace App\Tests\Functional\Services\ServiceStatusInspector;
 
 use App\Services\ServiceStatusInspector\ComponentInspectorInterface;
-use App\Services\ServiceStatusInspector\GenericServiceStatusInspector;
 use App\Services\ServiceStatusInspector\ServiceStatusInspector;
 use App\Tests\AbstractBaseFunctionalTest;
 use App\Tests\Services\HttpResponseFactory;
@@ -129,7 +128,7 @@ class ServiceStatusInspectorTest extends AbstractBaseFunctionalTest
         $componentInspectors = ObjectReflector::getProperty(
             $this->serviceStatusInspector,
             'componentInspectors',
-            GenericServiceStatusInspector::class
+            ServiceStatusInspector::class
         );
 
         if (array_key_exists($name, $componentInspectors)) {
@@ -138,7 +137,7 @@ class ServiceStatusInspectorTest extends AbstractBaseFunctionalTest
 
         ObjectReflector::setProperty(
             $this->serviceStatusInspector,
-            GenericServiceStatusInspector::class,
+            ServiceStatusInspector::class,
             'componentInspectors',
             $componentInspectors
         );
